backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.3
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.104.1
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
orjson==3.10.18
passlib==1.7.4
Pillow==10.1.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.5.1
pypng==0.20220715.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
qrcode==7.4.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
speedtest-cli==2.1.3
SQLAlchemy==2.0.23
starlette==0.27.0
typing-inspection==0.4.1
typing_extensions==4.13.2
ujson==5.10.0
uvicorn==0.24.0
watchfiles==1.0.5
websockets==15.0.1
asyncpg